chore(meshwide): move located links into a hook
